test: fix e2e tests
